Remove random seed reset inside main book loop
The seed pinning at the top of this block affects the global singleton
RNG instances, so it's not necessary to reset the seed inside the
main generation loop. Instead I think this was artificially
constraining the number of unique book outputs to 1000.
This change removes the inner seed setting, and also extracts a common
constant for NUM_BOOKS to be reused between mixture level generation
and book tour generation.